Revert "allow number of addresses to return -1 to indicate that the search just started for command line interface"
This reverts commit 61eeb11d24fe3a22ec18f2a65b8f5b6751aada28.

+++ src/kad.h
@@ -21,8 +21,7 @@ int kad_blacklist(const IP* addr);
 * Lookup the addresses of the nodes who have announced value id.
 * The first call will start the search.
 */
-
-int kad_lookup(const char query[], IP addr_array[], int *addr_num);
+int kad_lookup(const char query[], IP addr_array[], size_t *addr_num);
 
 // Export good nodes
 int kad_export_nodes(FILE *fp);
